redis-presence: release 0.18.4, fold the tests into the existing file
