+++ packages/presence/redis-presence/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@colyseus/redis-presence",
-  "version": "0.18.3",
+  "version": "0.18.4",
   "type": "module",
   "input": "./src/index.ts",
   "main": "./build/index.cjs",
